dual algorithm subclass algorithm
